refactor(container): extract lifecycle.go and credentials.go from container.go
Pure structural refactor — no behaviour changes. Splits the 1753-LoC
container.go into three focused files:

- lifecycle.go: EnsureRemoved, Create, WaitHealthy, Shutdown, and their
  private helpers (dumpLogs, hasExited, isHealthy).
- credentials.go: writeClaudeCredentials, credentialEnvVars, and the
  GitHub-URL parsing helpers (githubAccountFromBareRoot, githubAccountFromURL).
- container.go (residual): Manager struct, Config, constants, NameForSession,
  prepareVolumeDirs, buildRunArgs, writeGitconfig, writeSshConfig, and
  utility functions.

buildRunArgs remains in container.go as specified (deferred until A.1/A.2).

Closes #1154
