+++ modules/programs/prism/prism/internal/container/container.go
@@ -31,7 +31,6 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"runtime"
 	"strings"
 	"time"
 )
@@ -303,62 +302,7 @@ func New(cfg Config) *Manager {
 // Name returns the container name.
 func (m *Manager) Name() string { return m.name }
 
-// EnsureRemoved stops and removes any existing container with the same name,
-// and cleans up any temp files created by Create. It is safe to call when no
-// such container exists — errors for "no such container" are silently ignored.
-//
-// When the Manager was created with a non-empty Config.InstanceID, EnsureRemoved
-// first inspects the container's "prism.instance-id" label. If the label exists
-// and does not match the current InstanceID, a warning is logged indicating
-// that the container belongs to a different session instance. Removal still
-// proceeds regardless — the new session needs the container name.
-func (m *Manager) EnsureRemoved(ctx context.Context) {
-	// Clean up temp files created by Create.
-	_ = os.Remove(m.gitdirFilePath())
-	_ = os.Remove(m.worktreeGitdirFilePath())
-	_ = os.Remove(m.sshConfigFilePath())
-	_ = os.Remove(m.gitconfigFilePath())
-	_ = os.Remove(m.allowedSignersFilePath())
-	_ = os.Remove(m.opencodeConfigFilePath())
-	_ = os.Remove(m.claudeCredentialsFilePath())
-	_ = os.Remove(m.sandboxExecProfilePath())
 
-	// Check the container's instance label when we have our own InstanceID.
-	// This detects ownership mismatches where a container from a previous
-	// session incarnation is being cleaned up by a new one.
-	if m.cfg.InstanceID != "" {
-		inspectCtx, inspectCancel := context.WithTimeout(ctx, 5*time.Second)
-		out, inspectErr := exec.CommandContext(inspectCtx, "podman", "inspect",
-			"--format", `{{index .Config.Labels "prism.instance-id"}}`,
-			m.name,
-		).Output()
-		inspectCancel()
-		if inspectErr == nil {
-			containerInstanceID := strings.TrimSpace(string(out))
-			if containerInstanceID != "" && containerInstanceID != m.cfg.InstanceID {
-				log.Printf("container: warning: container %q has instance-id %q but current session has %q — removing anyway",
-					m.name, containerInstanceID, m.cfg.InstanceID)
-			}
-		}
-	}
-
-	// Stop the container (ignore errors — may not be running).
-	stopCmd := exec.CommandContext(ctx, "podman", "stop", "--time", "10", m.name)
-	if out, err := stopCmd.CombinedOutput(); err != nil {
-		// Only log if it looks like a real error (not "no such container").
-		if !IsNoSuchContainerError(string(out)) {
-			log.Printf("container: stop existing %q: %v — %s", m.name, err, strings.TrimSpace(string(out)))
-		}
-	}
-
-	// Remove the container (ignore errors — may not exist).
-	rmCmd := exec.CommandContext(ctx, "podman", "rm", "--force", m.name)
-	if out, err := rmCmd.CombinedOutput(); err != nil {
-		if !IsNoSuchContainerError(string(out)) {
-			log.Printf("container: rm existing %q: %v — %s", m.name, err, strings.TrimSpace(string(out)))
-		}
-	}
-}
 
 // sessionTempPath is the package-level building block for per-session temp
 // file paths. All per-session artefact files follow the shape:
@@ -457,43 +401,6 @@ func WriteOpencodeConfig(sessionName, content string) error {
 // the container where opencode-claude-auth can read it.
 func (m *Manager) claudeCredentialsFilePath() string {
 	return m.tempPath("claude-creds", "")
-}
-
-// writeClaudeCredentials extracts Claude Code credentials from the macOS
-// Keychain and writes them to a temp file. On Linux, Claude Code stores
-// credentials in ~/.claude/.credentials.json which is already inside the
-// claudeMount bind-mounted directory and requires no special handling. On
-// Darwin, credentials are stored in the Keychain under the service name
-// "Claude Code-credentials" and never reach disk, so the container never
-// sees them via the directory mount alone.
-//
-// Sets m.claudeCredentialsReady to true on success so that buildRunArgs can
-// add the bind-mount. Logs and returns without error on failure — a missing
-// Keychain entry (e.g. not logged in) should surface as an auth error from
-// opencode rather than a hard container launch failure.
-func (m *Manager) writeClaudeCredentials() {
-	m.claudeCredentialsReady = false
-	if runtime.GOOS != "darwin" {
-		return
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-	out, err := exec.CommandContext(ctx, "security", "find-generic-password",
-		"-l", "Claude Code-credentials", "-w").Output()
-	if err != nil {
-		log.Printf("container: could not extract Claude credentials from macOS Keychain: %v — opencode-claude-auth may fail to authenticate", err)
-		return
-	}
-	creds := strings.TrimSpace(string(out))
-	if creds == "" {
-		log.Printf("container: macOS Keychain returned empty Claude credentials — run `claude login` to authenticate")
-		return
-	}
-	if err := os.WriteFile(m.claudeCredentialsFilePath(), []byte(creds), 0o600); err != nil {
-		log.Printf("container: failed to write Claude credentials temp file: %v", err)
-		return
-	}
-	m.claudeCredentialsReady = true
 }
 
 // writeSshConfig generates a minimal ~/.ssh/config for the sandbox and writes
@@ -659,196 +566,7 @@ func (m *Manager) worktreeGitdirFilePath() string {
 // WorktreeGitdirFilePath is the exported version of worktreeGitdirFilePath for tests.
 func (m *Manager) WorktreeGitdirFilePath() string { return m.worktreeGitdirFilePath() }
 
-// Create creates and starts the podman container for this session.
-// It first calls EnsureRemoved to handle any stale container with the same name.
-// Returns an error if podman create or start fails.
-func (m *Manager) Create(ctx context.Context) error {
-	createStart := time.Now()
 
-	// Remove any stale container first (AC-15).
-	m.EnsureRemoved(ctx)
-	log.Printf("[timing] EnsureRemoved: %s", time.Since(createStart).Round(time.Millisecond))
-
-	// Write corrected git pointer files for the container.
-	if m.cfg.BareRoot != "" && m.cfg.WorktreeGitDir != "" {
-		branch := filepath.Base(m.cfg.WorktreeGitDir)
-
-		// Forward pointer: /workspace/.git → /prism-git/worktrees/<branch> (#492).
-		// The host's .git file contains an absolute host path that doesn't exist
-		// inside the container.
-		gitdirContent := "gitdir: /prism-git/worktrees/" + branch + "\n"
-		if err := os.WriteFile(m.gitdirFilePath(), []byte(gitdirContent), 0o644); err != nil {
-			return fmt.Errorf("container: write gitdir file: %w", err)
-		}
-
-		// Back-pointer: worktrees/<branch>/gitdir → /workspace/.git
-		// This file is the reverse pointer from git metadata back to the
-		// worktree checkout. On the host it contains the host worktree path
-		// (e.g. /home/user/code/repo/branch/.git) which doesn't exist in the
-		// container. nix/libgit2 follows this pointer when resolving the
-		// working tree, so it must point to the container-internal path.
-		wtGitdirContent := "/workspace/.git\n"
-		if err := os.WriteFile(m.worktreeGitdirFilePath(), []byte(wtGitdirContent), 0o644); err != nil {
-			return fmt.Errorf("container: write worktree gitdir file: %w", err)
-		}
-	}
-
-	// Write a minimal SSH config for the container. The host's ~/.ssh/config
-	// is a nix store symlink with wrong ownership that SSH rejects.
-	if err := m.writeSshConfig(isolationPodman); err != nil {
-		return err
-	}
-
-	// Write a minimal .gitconfig for the container. The host's git config lives
-	// in ~/.config/git/config (managed by home-manager) and is not mounted into
-	// containers. We generate a minimal config with identity, signing, and
-	// convenience settings.
-	t0 := time.Now()
-	if err := m.writeGitconfig(isolationPodman); err != nil {
-		return fmt.Errorf("container: write gitconfig: %w", err)
-	}
-	log.Printf("[timing] writeGitconfig: %s", time.Since(t0).Round(time.Millisecond))
-
-	// Write the opencode config file for the container. This replaces the
-	// previous OPENCODE_CONFIG_CONTENT env var approach so that relative plugin
-	// paths (e.g. "./plugins/my-plugin") resolve correctly from the config
-	// file's location at /root/.config/opencode/opencode.json.
-	if m.cfg.ConfigContent != "" {
-		if err := os.WriteFile(m.opencodeConfigFilePath(), []byte(m.cfg.ConfigContent), 0o644); err != nil {
-			return fmt.Errorf("container: write opencode config: %w", err)
-		}
-	}
-
-	// On Darwin, extract Claude Code credentials from the macOS Keychain and
-	// write them to a temp file so the container can find them. On Linux,
-	// ~/.claude/.credentials.json is already inside the claudeMount directory.
-	m.writeClaudeCredentials()
-
-	// Pre-create directories that buildRunArgs() will reference as volume mount
-	// sources. podman validates all bind-mount source paths before starting the
-	// container and exits 125 if any are absent, so we create them eagerly here.
-	// Failures are logged and non-fatal — podman will surface the real error if
-	// the directory is still absent after this attempt.
-	if err := m.prepareVolumeDirs(true); err != nil {
-		// prepareVolumeDirs logs individual failures; a non-nil error means
-		// multiple dirs failed and is logged here for observability only.
-		log.Printf("container: prepareVolumeDirs partial failure: %v", err)
-	}
-
-	// Build the podman run arguments.
-	t0 = time.Now()
-	args := m.buildRunArgs()
-	log.Printf("[timing] buildRunArgs: %s", time.Since(t0).Round(time.Millisecond))
-
-	log.Printf("container: creating %q: podman %s", m.name, strings.Join(redactArgs(args), " "))
-
-	podmanStart := time.Now()
-	if err := m.isolator.Run(ctx, args); err != nil {
-		return err
-	}
-	log.Printf("[timing] podman run: %s (total to here: %s)",
-		time.Since(podmanStart).Round(time.Millisecond),
-		time.Since(createStart).Round(time.Millisecond))
-	log.Printf("[timing] Create total: %s", time.Since(createStart).Round(time.Millisecond))
-
-	return nil
-}
-
-// WaitHealthy polls the container's HTTP endpoint until it responds with a
-// successful status code (2xx) or the timeout expires. Returns nil when the
-// container is healthy.
-func (m *Manager) WaitHealthy(ctx context.Context) error {
-	timeout := m.cfg.HealthCheckTimeout
-	if timeout == 0 {
-		timeout = DefaultHealthCheckTimeout
-	}
-
-	waitStart := time.Now()
-	probeCount := 0
-	deadline := waitStart.Add(timeout)
-	for {
-		if time.Now().After(deadline) {
-			m.dumpLogs()
-			return fmt.Errorf("container: health check timed out after %s for %q", timeout, m.name)
-		}
-		if ctx.Err() != nil {
-			m.dumpLogs()
-			return ctx.Err()
-		}
-
-		probeCount++
-		healthy := m.isHealthy(ctx)
-		elapsed := time.Since(waitStart).Round(time.Millisecond)
-		if healthy {
-			log.Printf("[timing] WaitHealthy probe %d: %s elapsed — ok", probeCount, elapsed)
-			log.Printf("[timing] WaitHealthy: %s (%d probes)", time.Since(waitStart).Round(time.Millisecond), probeCount)
-			return nil
-		}
-		log.Printf("[timing] WaitHealthy probe %d: %s elapsed — fail", probeCount, elapsed)
-
-		// Check if the container exited early — no point polling a dead
-		// container until the timeout expires.
-		if exited, code := m.hasExited(); exited {
-			m.dumpLogs()
-			return fmt.Errorf("container: %q exited with code %d before becoming healthy", m.name, code)
-		}
-
-		select {
-		case <-ctx.Done():
-			m.dumpLogs()
-			return ctx.Err()
-		case <-time.After(healthCheckInterval):
-		}
-	}
-}
-
-// dumpLogs writes the container's stdout/stderr to the sidecar log so that
-// startup failures are visible without needing to race `podman logs` before
-// the container is removed by Shutdown.
-func (m *Manager) dumpLogs() {
-	m.isolator.DumpLogs()
-}
-
-// hasExited checks whether the container has already stopped. Returns true and
-// the exit code if the container is in an exited state.
-func (m *Manager) hasExited() (bool, int) {
-	return m.isolator.HasExited()
-}
-
-// isHealthy performs a single HTTP probe and returns true on success.
-func (m *Manager) isHealthy(ctx context.Context) bool {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, m.healthCheckURL, nil)
-	if err != nil {
-		return false
-	}
-	resp, err := m.httpClient.Do(req)
-	if err != nil {
-		return false
-	}
-	_ = resp.Body.Close()
-	return resp.StatusCode >= 200 && resp.StatusCode < 300
-}
-
-// Shutdown stops and removes the container. Intended to be called on SIGTERM.
-// It delegates the podman stop/rm calls to the Isolator and then cleans up
-// the temp files created by Create.
-func (m *Manager) Shutdown() {
-	log.Printf("container: shutting down %q", m.name)
-
-	m.isolator.Shutdown()
-
-	// Clean up temp files created by Create.
-	_ = os.Remove(m.gitdirFilePath())
-	_ = os.Remove(m.worktreeGitdirFilePath())
-	_ = os.Remove(m.sshConfigFilePath())
-	_ = os.Remove(m.gitconfigFilePath())
-	_ = os.Remove(m.allowedSignersFilePath())
-	_ = os.Remove(m.opencodeConfigFilePath())
-	_ = os.Remove(m.claudeCredentialsFilePath())
-	_ = os.Remove(m.sandboxExecProfilePath())
-
-	log.Printf("container: %q removed", m.name)
-}
 
 // PrepareBwrap writes the same temp files that Create() writes for podman
 // sessions (SSH config, gitconfig, opencode.json config) and returns the
@@ -1535,136 +1253,7 @@ func (m *Manager) buildRunArgs() []string {
 	return args
 }
 
-// githubAccountFromBareRoot returns the GitHub account (organisation or user)
-// for the repo by reading the origin remote URL from the bare git dir.
-// Returns "" when the bare root is empty, git is unavailable, or the remote
-// URL does not match a github.com URL pattern.
-//
-// Supported URL formats:
-//
-//	git@github.com:<account>/<repo>.git   (SSH)
-//	https://github.com/<account>/<repo>   (HTTPS, with or without .git)
-func githubAccountFromBareRoot(bareRoot string) string {
-	if bareRoot == "" {
-		return ""
-	}
-	// The bare git dir lives at <bareRoot>/.bare — use --git-dir to run git
-	// against it directly without needing to be inside a worktree.
-	bareDir := filepath.Join(bareRoot, ".bare")
-	cmd := exec.Command("git", "--git-dir", bareDir, "remote", "get-url", "origin")
-	out, err := cmd.Output()
-	if err != nil {
-		// Also try the bareRoot itself in case it IS the raw bare git dir.
-		cmd2 := exec.Command("git", "--git-dir", bareRoot, "remote", "get-url", "origin")
-		out, err = cmd2.Output()
-		if err != nil {
-			return ""
-		}
-	}
-	return githubAccountFromURL(strings.TrimSpace(string(out)))
-}
 
-// githubAccountFromURL parses a git remote URL and returns the GitHub account
-// (the path segment immediately after "github.com"). Returns "" if the URL is
-// not a recognisable github.com URL.
-func githubAccountFromURL(remoteURL string) string {
-	// SSH: git@github.com:<account>/<repo>[.git]
-	if strings.HasPrefix(remoteURL, "git@github.com:") {
-		rest := strings.TrimPrefix(remoteURL, "git@github.com:")
-		parts := strings.SplitN(rest, "/", 2)
-		if len(parts) >= 1 && parts[0] != "" {
-			return parts[0]
-		}
-	}
-	// HTTPS: https://github.com/<account>/<repo>[.git]
-	//        https://x-access-token:TOKEN@github.com/<account>/<repo>[.git]
-	if idx := strings.Index(remoteURL, "github.com/"); idx >= 0 {
-		rest := remoteURL[idx+len("github.com/"):]
-		parts := strings.SplitN(rest, "/", 2)
-		if len(parts) >= 1 && parts[0] != "" {
-			return parts[0]
-		}
-	}
-	return ""
-}
-
-// credentialEnvVars returns the environment variable assignments to inject into
-// the container based on the agent role and current host environment.
-// Only vars that are set on the host are forwarded — unset vars are skipped.
-//
-// GitHub token selection (4-PAT architecture):
-// The correct token is chosen based on the GitHub account (derived from the
-// repo's origin remote URL) and the agent role:
-//
-//	PRISM_GITHUB_TOKEN_PRISMATIC_KOI_COORDINATOR   — prismatic-koi + coordinator
-//	PRISM_GITHUB_TOKEN_PRISMATIC_KOI_WORKER        — prismatic-koi + worker
-//	PRISM_GITHUB_TOKEN_THANKYOU_PAYROLL_COORDINATOR — thankyou-payroll + coordinator
-//	PRISM_GITHUB_TOKEN_THANKYOU_PAYROLL_WORKER      — thankyou-payroll + worker
-//
-// Falls back to host GITHUB_TOKEN if the specific token is not set
-// (supports host-mode, --host-mode spawns, and migration period).
-func (m *Manager) credentialEnvVars() []string {
-	var vars []string
-
-	// LLM API keys — forwarded for all agent roles.
-	llmKeys := []string{
-		"ANTHROPIC_API_KEY",
-		"OPENAI_API_KEY",
-		"GEMINI_API_KEY",
-		"GOOGLE_API_KEY",
-		"GITHUB_COPILOT_TOKEN",
-		"DEEPSEEK_API_KEY",
-		"OPENROUTER_API_KEY",
-	}
-	for _, k := range llmKeys {
-		if v := os.Getenv(k); v != "" {
-			vars = append(vars, k+"="+v)
-		}
-	}
-
-	// GitHub token — 4-PAT architecture: account × role → specific token.
-	// Derive the account from the repo's git remote URL.
-	account := githubAccountFromBareRoot(m.cfg.BareRoot)
-	role := m.cfg.AgentRole
-
-	// Build the env var name: PRISM_GITHUB_TOKEN_<ACCOUNT>_<ROLE>
-	// where account is uppercased with hyphens replaced by underscores.
-	var tokenEnvVar string
-	if account != "" {
-		accountKey := strings.ToUpper(strings.ReplaceAll(account, "-", "_"))
-		roleKey := strings.ToUpper(role)
-		if roleKey == "WORKER" || roleKey == "COORDINATOR" {
-			tokenEnvVar = "PRISM_GITHUB_TOKEN_" + accountKey + "_" + roleKey
-		}
-	}
-
-	// Try specific token first, then fall back to host GITHUB_TOKEN.
-	if tokenEnvVar != "" {
-		if tok := os.Getenv(tokenEnvVar); tok != "" {
-			vars = append(vars, "GITHUB_TOKEN="+tok)
-			return vars
-		}
-	}
-	// Fallback: use host GITHUB_TOKEN (supports --host-mode and migration period).
-	if tok := os.Getenv("GITHUB_TOKEN"); tok != "" {
-		vars = append(vars, "GITHUB_TOKEN="+tok)
-	}
-
-	// Note: GIT_AUTHOR_NAME, GIT_AUTHOR_EMAIL, GIT_COMMITTER_NAME, and
-	// GIT_COMMITTER_EMAIL are intentionally NOT injected. The container now
-	// has a generated .gitconfig with a [user] section (sourced from prism
-	// config). Env vars override gitconfig and would mask a broken gitconfig.
-
-	// Note: GIT_DIR and GIT_COMMON_DIR are intentionally NOT injected.
-	// Instead, Create() writes a corrected .git pointer file and bind-mounts it
-	// over /workspace/.git so all tools — including opencode's internal git
-	// library which reads .git directly rather than honouring GIT_DIR — resolve
-	// the correct container-internal path (#492).
-	// GIT_COMMON_DIR breaks ref lookup in the git version used in the container
-	// image and is therefore also omitted.
-
-	return vars
-}
 
 // redactArgs returns a copy of args where any value immediately following a
 // "--env" element has its value (the part after the first "=") replaced with
